+++ R/SimInf_model.R
@@ -4,7 +4,7 @@
 ## Copyright (C) 2015 Pavol Bauer
 ## Copyright (C) 2017 -- 2019 Robin Eriksson
 ## Copyright (C) 2015 -- 2019 Stefan Engblom
-## Copyright (C) 2015 -- 2020 Stefan Widgren
+## Copyright (C) 2015 -- 2021 Stefan Widgren
 ##
 ## SimInf is free software: you can redistribute it and/or modify
 ## it under the terms of the GNU General Public License as published by
@@ -19,54 +19,10 @@
 ## You should have received a copy of the GNU General Public License
 ## along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-##' Class \code{"SimInf_model"}
-##'
-##' Class to handle data for the \code{SimInf_model}.
-##' @template G-slot
-##' @template S-slot
-##' @template U-slot
-##' @template U_sparse-slot
-##' @slot V The result matrix for the real-valued continuous
-##'     state. \code{V[, j]} contains the real-valued state of the
-##'     system at \code{tspan[j]}. Numeric matrix
-##'     (\eqn{N_n}\code{dim(ldata)[1]} \eqn{\times}
-##'     \code{length(tspan)}).
-##' @slot V_sparse If the model was configured to write the solution
-##'     to a sparse matrix (\code{dgCMatrix}) the \code{V_sparse}
-##'     contains the data and \code{V} is empty. The layout of the
-##'     data in \code{V_sparse} is identical to \code{V}.
-##' @template ldata-slot
-##' @template gdata-slot
-##' @template tspan-slot
-##' @template u0-slot
-##' @slot v0 The initial value for the real-valued continuous state.
-##'     Numeric matrix (\code{dim(ldata)[1]} \eqn{\times N_n}).
-##' @slot events Scheduled events \code{\linkS4class{SimInf_events}}
-##' @template C_code-slot
-##' @include SimInf_events.R
-##' @export
-##' @importFrom methods validObject
-##' @importClassesFrom Matrix dgCMatrix
-setClass(
-    "SimInf_model",
-    slots = c(G        = "dgCMatrix",
-              S        = "dgCMatrix",
-              U        = "matrix",
-              U_sparse = "dgCMatrix",
-              ldata    = "matrix",
-              gdata    = "numeric",
-              tspan    = "numeric",
-              u0       = "matrix",
-              V        = "matrix",
-              V_sparse = "dgCMatrix",
-              v0       = "matrix",
-              events   = "SimInf_events",
-              C_code   = "character")
-)
-
 ##' Check if a SimInf_model object is valid
 ##'
 ##' @param object The SimInf_model_object to check.
+##' @include classes.R
 ##' @include valid.R
 ##' @noRd
 valid_SimInf_model_object <- function(object) {
@@ -188,10 +144,23 @@ SimInf_model <- function(G,
 ##'
 ##' ## Extract the global data vector that is common to all nodes
 ##' gdata(model)
-gdata <- function(model) {
-    check_model_argument(model)
-    model@gdata
-}
+setGeneric(
+    "gdata",
+    signature = "model",
+    function(model) {
+        standardGeneric("gdata")
+    }
+)
+
+##' @rdname gdata
+##' @export
+setMethod(
+    "gdata",
+    signature(model = "SimInf_model"),
+    function(model) {
+        model@gdata
+    }
+)
 
 ##' Set a global data parameter for a \code{SimInf_model} object
 ##'
@@ -213,25 +182,37 @@ gdata <- function(model) {
 ##'
 ##' ## Extract the global data vector that is common to all nodes
 ##' gdata(model)
-"gdata<-" <- function(model, parameter, value) {
-    check_model_argument(model)
+setGeneric(
+    "gdata<-",
+    signature = "model",
+    function(model, parameter, value) {
+        standardGeneric("gdata<-")
+    }
+)
 
-    ## Check paramter argument
-    if (missing(parameter))
-        stop("Missing 'parameter' argument.", call. = FALSE)
-    if (!is.character(parameter))
-        stop("'parameter' argument must be a character.", call. = FALSE)
+##' @rdname gdata-set
+##' @export
+setMethod(
+    "gdata<-",
+    signature(model = "SimInf_model"),
+    function(model, parameter, value) {
+        ## Check parameter argument
+        if (missing(parameter))
+            stop("Missing 'parameter' argument.", call. = FALSE)
+        if (!is.character(parameter))
+            stop("'parameter' argument must be a character.", call. = FALSE)
 
-    ## Check value argument
-    if (missing(value))
-        stop("Missing 'value' argument.", call. = FALSE)
-    if (!is.numeric(value))
-        stop("'value' argument must be a numeric.", call. = FALSE)
+        ## Check value argument
+        if (missing(value))
+            stop("Missing 'value' argument.", call. = FALSE)
+        if (!is.numeric(value))
+            stop("'value' argument must be a numeric.", call. = FALSE)
 
-    model@gdata[parameter] <- value
+        model@gdata[parameter] <- value
 
-    model
-}
+        model
+    }
+)
 
 ##' Extract local data from a node
 ##'
@@ -253,14 +234,26 @@ gdata <- function(model) {
 ##' ## Display local data from the first two nodes.
 ##' ldata(model, node = 1)
 ##' ldata(model, node = 2)
-ldata <- function(model, node) {
-    check_model_argument(model)
+setGeneric(
+    "ldata",
+    signature = "model",
+    function(model, node) {
+        standardGeneric("ldata")
+    }
+)
 
-    ## Check node argument
-    if (missing(node))
-        stop("Missing 'node' argument.", call. = FALSE)
-    if (!is.numeric(node) || !identical(length(node), 1L) || node < 1)
-        stop("Invalid 'node' argument.", call. = FALSE)
+##' @rdname ldata
+##' @export
+setMethod(
+    "ldata",
+    signature(model = "SimInf_model"),
+    function(model, node) {
+        ## Check node argument
+        if (missing(node))
+            stop("Missing 'node' argument.", call. = FALSE)
+        if (!is.numeric(node) || !identical(length(node), 1L) || node < 1)
+            stop("Invalid 'node' argument.", call. = FALSE)
 
-    model@ldata[, node]
-}
+        model@ldata[, node]
+    }
+)